refactor: move history height logic to hook
